+++ src/lib/packages/calendar/ui/src/lib/context.ts
@@ -1,0 +1,12 @@
+import { Context } from "@cluue/utils";
+import type { createCalendarStore } from "./store.js";
+
+interface ICalendarContextData {
+    store:ReturnType<typeof createCalendarStore>
+}
+
+export class CalendarContext extends Context<ICalendarContextData> {
+    constructor() {
+        super('calendar')
+    }
+}