+++ src/lib/packages/calendar/ui/CHANGELOG.md
@@ -4,10 +4,10 @@
 
 ### Patch Changes
 
-- Fix test
+-   Fix test
 
 ## 1.0.1
 
 ### Patch Changes
 
-- Create a calendar packages
+-   Create a calendar packages